requirements.txt
easyocr==1.6.2
opencv-python-headless==4.5.4.60
pillow==9.2.0
numpy
requests==2.18.4
langid==1.1.5
